src/main/java/com/ncr/chess/Pawn.java
package com.ncr.chess;

import java.awt.*;
import java.util.ArrayList;
import java.util.Collection;
import java.util.List;

public class Pawn {

    private ChessBoard chessBoard;
    private int xCoordinate;
    private int yCoordinate;
    private PieceColor pieceColor;
    private Coordinate coordinate;

    public Pawn(PieceColor pieceColor) {
        this.pieceColor = pieceColor;
    }

    public ChessBoard getChessBoard() {
        return chessBoard;
    }

    public void setChessBoard(ChessBoard chessBoard) {
        this.chessBoard = chessBoard;
    }

    public int getXCoordinate() {
        return xCoordinate;
    }

    public void setXCoordinate(int value) {
        this.xCoordinate = value;
    }

    public int getYCoordinate() {
        return yCoordinate;
    }

    public void setYCoordinate(int value) {
        this.yCoordinate = value;
    }

    public PieceColor getPieceColor() {
        return this.pieceColor;
    }

    private void setPieceColor(PieceColor value) {
        pieceColor = value;
    }

    public Coordinate getCoordinate() {
        return coordinate;
    }

    public void setCoordinate(Coordinate coordinate) {
        this.coordinate = coordinate;
    }


    public void move(MovementType movementType, Coordinate sourcCoordinates,Coordinate destCoordinates,PieceColor color) throws UnsupportedOperationException {
        try {
            if(getChessBoard().isLegalBoardPosition(sourcCoordinates.positionX,sourcCoordinates.positionY,coordinate,this))
            {
                System.out.println("Moved the PAWN");
            }
        } catch (Exception e) {
            throw new UnsupportedOperationException("Need to implement Pawn.Move()");
        }

    }

    @Override
    public String toString() {
        return getCurrentPositionAsString();
    }

    protected String getCurrentPositionAsString() {
        String eol = System.lineSeparator();
        return String.format("Current X: {1}{0}Current Y: {2}{0}Piece Color: {3}", eol, xCoordinate, yCoordinate, pieceColor);
    }
}
